86d015k33 fixed architectural issues and feedbacks from the peer review
